refactor(app): extract set_active_buffer + focus_split (phase 6g)
Move set_active_buffer and focus_split out of mod.rs into a new
app/active_focus.rs sibling. ~196 lines.

These are the centralized methods for switching what the user is
looking at — split-manager updates, tab list updates, file-explorer
sync, terminal buffer resume, semantic-token cleanup. Visibility
opens to pub(super) so all the call sites in extracted phases
(buffer_close, file_open_orchestrators, etc.) keep working.

mod.rs goes from 3485 -> 3289 lines (-196). 2026 tests still pass.
